scripts/blog_generator/requirements.txt
feedgen==1.0
click==8.1
pyyaml==6.0